Finished login and updated visuals and more
Finished the login feature; connected the activities and other small features for instance the Keyboard now disappears when user clicks on the screen around the other elements. NOTE: the username and password are still a string and not stored in a Json. Also, progress bar are not done yet, still having some issues there.